+++ src/main/java/com/buuz135/industrial/gui/component/FilterGuiComponent.java
@@ -1,5 +1,6 @@
 package com.buuz135.industrial.gui.component;
 
+import com.buuz135.industrial.api.conveyor.gui.PositionedGuiComponent;
 import com.buuz135.industrial.gui.conveyor.GuiConveyor;
 import com.buuz135.industrial.proxy.block.filter.IFilter;
 import com.buuz135.industrial.utils.Reference;
