+++ src/main/java/com/buuz135/industrial/gui/component/TextGuiComponent.java
@@ -1,6 +1,7 @@
 package com.buuz135.industrial.gui.component;
 
 
+import com.buuz135.industrial.api.conveyor.gui.PositionedGuiComponent;
 import com.buuz135.industrial.gui.conveyor.GuiConveyor;
 import net.minecraft.client.Minecraft;
