+++ src/main/java/com/buuz135/industrial/gui/component/TexturedStateButtonGuiComponent.java
@@ -1,5 +1,6 @@
 package com.buuz135.industrial.gui.component;
 
+import com.buuz135.industrial.api.conveyor.gui.PositionedGuiComponent;
 import com.buuz135.industrial.gui.conveyor.GuiConveyor;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.audio.PositionedSoundRecord;
