+++ src/main/java/com/buuz135/industrial/gui/conveyor/GuiConveyor.java
@@ -2,8 +2,8 @@ package com.buuz135.industrial.gui.conveyor;
 
 import com.buuz135.industrial.IndustrialForegoing;
 import com.buuz135.industrial.api.conveyor.ConveyorUpgrade;
+import com.buuz135.industrial.api.conveyor.gui.IGuiComponent;
 import com.buuz135.industrial.gui.component.FilterGuiComponent;
-import com.buuz135.industrial.gui.component.IGuiComponent;
 import com.buuz135.industrial.proxy.block.filter.IFilter;
 import com.buuz135.industrial.proxy.network.ConveyorButtonInteractMessage;
 import com.buuz135.industrial.utils.Reference;
