+++ src/main/java/com/buuz135/industrial/proxy/block/ConveyorDetectorUpgrade.java
@@ -3,8 +3,8 @@ package com.buuz135.industrial.proxy.block;
 import com.buuz135.industrial.api.conveyor.ConveyorUpgrade;
 import com.buuz135.industrial.api.conveyor.ConveyorUpgradeFactory;
 import com.buuz135.industrial.api.conveyor.IConveyorContainer;
+import com.buuz135.industrial.api.conveyor.gui.IGuiComponent;
 import com.buuz135.industrial.gui.component.FilterGuiComponent;
-import com.buuz135.industrial.gui.component.IGuiComponent;
 import com.buuz135.industrial.gui.component.StateButtonInfo;
 import com.buuz135.industrial.gui.component.TexturedStateButtonGuiComponent;
 import com.buuz135.industrial.proxy.block.filter.IFilter;
